back_end3.0/requirements.txt
appdirs==1.4.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cors==0.9.7
cvxopt==1.2.5.post1
cycler==0.10.0
distlib==0.3.1
document==1.0
filelock==3.0.12
finance==0.2502
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
jqdatasdk==1.8.6
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
msgpack==1.0.2
numpy==1.19.5
pandas==0.25.3
Pillow==8.1.0
ply==3.11
progressbar==2.5
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.5
requests==2.25.1
requests-file==1.5.1
scipy==1.6.0
six==1.15.0
SQLAlchemy==1.3.22
thriftpy2==0.4.14
tldextract==3.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.0
Werkzeug==1.0.1
xlrd==1.2.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
